LayoutTests/platform/chromium/fast/forms/suggestion-picker/time-suggestion-picker-step-attribute.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../../../fast/js/resources/js-test-pre.js"></script>
<script src="../../../../../fast/forms/resources/picker-common.js"></script>
<script src="resources/suggestion-picker-common.js"></script>
</head>
<body style="background-color: #bbbbbb;">
<p id="description"></p>
<div id="console"></div>
<input type=time id=time value="00:00" step="3600" list=suggestions>
<datalist id=suggestions>
    <option>01:00</option>
    <option>01:00:01</option>
    <option>01:00:01.001</option>
    <option>01:00:01.010</option>
    <option>01:01</option>
    <option>02:00</option>
    <option>02:01</option>
</datalist>
<script>
description('Tests that step attribute filter out suggestions as expected.');

debug('Check that page popup doesn\'t exist at first.');
shouldBeNull('document.getElementById("mock-page-popup")');

var timeElement = document.getElementById('time');

window.onload = function() {
    openPicker(timeElement, test1);
};

function test1() {
    popupWindow.focus();

    shouldBe('entryValues().length', '2');
    shouldBeEqualToString('entryValues()[0]', '01:00');
    shouldBeEqualToString('entryValues()[1]', '02:00');

    eventSender.keyDown('\x1B'); // Close picker.
    waitUntilClosing(test1AfterClosing);
}

function test1AfterClosing() {
    timeElement.step = 60;
    openPicker(timeElement, test2);
}

function test2() {
    popupWindow.focus();

    shouldBe('entryValues().length', '4');
    shouldBeEqualToString('entryValues()[0]', '01:00');
    shouldBeEqualToString('entryValues()[1]', '01:01');
    shouldBeEqualToString('entryValues()[2]', '02:00');
    shouldBeEqualToString('entryValues()[3]', '02:01');

    eventSender.keyDown('\x1B'); // Close picker.
    waitUntilClosing(test2AfterClosing);
}

function test2AfterClosing() {
    timeElement.step = 1;
    openPicker(timeElement, test3);
}

function test3() {
    popupWindow.focus();

    shouldBe('entryValues().length', '5');
    shouldBeEqualToString('entryValues()[0]', '01:00');
    shouldBeEqualToString('entryValues()[1]', '01:00:01');
    shouldBeEqualToString('entryValues()[2]', '01:01');
    shouldBeEqualToString('entryValues()[3]', '02:00');
    shouldBeEqualToString('entryValues()[4]', '02:01');

    eventSender.keyDown('\x1B'); // Close picker.
    waitUntilClosing(test3AfterClosing);
}

function test3AfterClosing() {
    timeElement.step = 0.001;
    openPicker(timeElement, test4);
}

function test4() {
    popupWindow.focus();

    shouldBe('entryValues().length', '7');
    shouldBeEqualToString('entryValues()[0]', '01:00');
    shouldBeEqualToString('entryValues()[1]', '01:00:01');
    shouldBeEqualToString('entryValues()[2]', '01:00:01.001');
    shouldBeEqualToString('entryValues()[3]', '01:00:01.010');
    shouldBeEqualToString('entryValues()[4]', '01:01');
    shouldBeEqualToString('entryValues()[5]', '02:00');
    shouldBeEqualToString('entryValues()[6]', '02:01');

    finishJSTest();
}

</script>
<script src="../../../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
