LayoutTests/platform/chromium/fast/forms/suggestion-picker/time-suggestion-picker-step-attribute-expected.txt

Tests that step attribute filter out suggestions as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Check that page popup doesn't exist at first.
PASS document.getElementById("mock-page-popup") is null
PASS entryValues().length is 2
PASS entryValues()[0] is "01:00"
PASS entryValues()[1] is "02:00"
PASS entryValues().length is 4
PASS entryValues()[0] is "01:00"
PASS entryValues()[1] is "01:01"
PASS entryValues()[2] is "02:00"
PASS entryValues()[3] is "02:01"
PASS entryValues().length is 5
PASS entryValues()[0] is "01:00"
PASS entryValues()[1] is "01:00:01"
PASS entryValues()[2] is "01:01"
PASS entryValues()[3] is "02:00"
PASS entryValues()[4] is "02:01"
PASS entryValues().length is 7
PASS entryValues()[0] is "01:00"
PASS entryValues()[1] is "01:00:01"
PASS entryValues()[2] is "01:00:01.001"
PASS entryValues()[3] is "01:00:01.010"
PASS entryValues()[4] is "01:01"
PASS entryValues()[5] is "02:00"
PASS entryValues()[6] is "02:01"
PASS successfullyParsed is true

TEST COMPLETE
 
